+++ test/on.cpp
@@ -1,10 +1,10 @@
 #include "detail/common.hpp"
 
 #include <async/concepts.hpp>
-#include <async/inline_scheduler.hpp>
 #include <async/just.hpp>
 #include <async/just_result_of.hpp>
 #include <async/on.hpp>
+#include <async/schedulers/inline_scheduler.hpp>
 #include <async/stop_token.hpp>
 #include <async/when_all.hpp>
