+++ test/retry.cpp
@@ -1,11 +1,11 @@
 #include "detail/common.hpp"
 
 #include <async/concepts.hpp>
-#include <async/inline_scheduler.hpp>
 #include <async/just.hpp>
 #include <async/let_value.hpp>
 #include <async/on.hpp>
 #include <async/retry.hpp>
+#include <async/schedulers/inline_scheduler.hpp>
 #include <async/variant_sender.hpp>
 #include <async/when_all.hpp>
