+++ test/schedulers/thread_scheduler.cpp
@@ -3,7 +3,7 @@
 #include <async/completion_scheduler.hpp>
 #include <async/concepts.hpp>
 #include <async/env.hpp>
-#include <async/thread_scheduler.hpp>
+#include <async/schedulers/thread_scheduler.hpp>
 
 #include <catch2/catch_test_macros.hpp>
