+++ test/then.cpp
@@ -2,8 +2,8 @@
 
 #include <async/concepts.hpp>
 #include <async/env.hpp>
-#include <async/inline_scheduler.hpp>
 #include <async/just.hpp>
+#include <async/schedulers/inline_scheduler.hpp>
 #include <async/then.hpp>
 
 #include <catch2/catch_test_macros.hpp>
