+++ test/when_any.cpp
@@ -2,9 +2,9 @@
 
 #include <async/concepts.hpp>
 #include <async/just.hpp>
+#include <async/schedulers/thread_scheduler.hpp>
 #include <async/sync_wait.hpp>
 #include <async/then.hpp>
-#include <async/thread_scheduler.hpp>
 #include <async/type_traits.hpp>
 #include <async/when_any.hpp>
